+++ models/clienttel.go
@@ -3,8 +3,8 @@ package models
 //ClientTel numero telefonico del cliente
 type ClientTel struct {
 	ModelBig
-	CodClient     int        `json:"id_cliente,omit" gorm:"integer"`
-	Phone         int        `json:"number,omitempty" gorm:"NUMERIC"`
+	CodClient     uint32     `json:"id_cliente,omit" gorm:"integer"`
+	Phone         uint64     `json:"number,omitempty" gorm:"NUMERIC"`
 	CodTelDescrip int8       `json:"id_tel_descript,omit" gorm:"SMALLINT"`
 	TelDescrip    TelDescrip `json:"type_tel,omitempty"`
 }
